+++ src/app.rs
@@ -2,10 +2,10 @@ use crate::{
 	accessors,
 	cmdbar::CommandBar,
 	components::{
-		event_pump, AppOption, BlameFileComponent,
-		BranchListComponent, CommandBlocking, CommandInfo,
-		CommitComponent, CompareCommitsComponent, Component,
-		ConfirmComponent, CreateBranchComponent, DrawableComponent,
+		command_pump, event_pump, AppOption, BlameFileComponent,
+		BranchListComponent, CommandInfo, CommitComponent,
+		CompareCommitsComponent, Component, ConfirmComponent,
+		CreateBranchComponent, DrawableComponent,
 		ExternalEditorComponent, FetchComponent, FileRevlogComponent,
 		FuzzyFindPopup, FuzzyFinderTarget, HelpComponent,
 		InspectCommitComponent, LogSearchPopupComponent,
@@ -20,7 +20,8 @@ use crate::{
 	options::{Options, SharedOptions},
 	popup_stack::PopupStack,
 	queue::{
-		Action, InternalEvent, NeedsUpdate, Queue, StackablePopupOpen,
+		Action, AppTabs, InternalEvent, NeedsUpdate, Queue,
+		StackablePopupOpen,
 	},
 	setup_popups,
 	strings::{self, ellipsis_trim_start, order},
@@ -273,6 +274,7 @@ impl App {
 				key_config.clone(),
 			),
 			log_search_popup: LogSearchPopupComponent::new(
+				repo.clone(),
 				&queue,
 				theme.clone(),
 				key_config.clone(),
@@ -696,15 +698,15 @@ impl App {
 
 	fn switch_tab(&mut self, k: &KeyEvent) -> Result<()> {
 		if key_match(k, self.key_config.keys.tab_status) {
-			self.set_tab(0)?;
+			self.switch_to_tab(&AppTabs::Status)?;
 		} else if key_match(k, self.key_config.keys.tab_log) {
-			self.set_tab(1)?;
+			self.switch_to_tab(&AppTabs::Log)?;
 		} else if key_match(k, self.key_config.keys.tab_files) {
-			self.set_tab(2)?;
+			self.switch_to_tab(&AppTabs::Files)?;
 		} else if key_match(k, self.key_config.keys.tab_stashing) {
-			self.set_tab(3)?;
+			self.switch_to_tab(&AppTabs::Stashing)?;
 		} else if key_match(k, self.key_config.keys.tab_stashes) {
-			self.set_tab(4)?;
+			self.switch_to_tab(&AppTabs::Stashlist)?;
 		}
 
 		Ok(())
@@ -723,6 +725,17 @@ impl App {
 		self.tab = tab;
 		self.options.borrow_mut().set_current_tab(tab);
 
+		Ok(())
+	}
+
+	fn switch_to_tab(&mut self, tab: &AppTabs) -> Result<()> {
+		match tab {
+			AppTabs::Status => self.set_tab(0)?,
+			AppTabs::Log => self.set_tab(1)?,
+			AppTabs::Files => self.set_tab(2)?,
+			AppTabs::Stashing => self.set_tab(3)?,
+			AppTabs::Stashlist => self.set_tab(4)?,
+		}
 		Ok(())
 	}
 
@@ -854,6 +867,10 @@ impl App {
 				self.tags_popup.open()?;
 			}
 			InternalEvent::TabSwitchStatus => self.set_tab(0)?,
+			InternalEvent::TabSwitch(tab) => {
+				self.switch_to_tab(&tab)?;
+				flags.insert(NeedsUpdate::ALL);
+			}
 			InternalEvent::SelectCommitInRevlog(id) => {
 				if let Err(error) = self.revlog.select_commit(id) {
 					self.queue.push(InternalEvent::ShowErrorMsg(
@@ -975,7 +992,7 @@ impl App {
 				self.reset_popup.open(id)?;
 			}
 			InternalEvent::CommitSearch(options) => {
-				self.revlog.search(options)?;
+				self.revlog.search(options);
 			}
 		};
 
@@ -1116,14 +1133,7 @@ impl App {
 	fn commands(&self, force_all: bool) -> Vec<CommandInfo> {
 		let mut res = Vec::new();
 
-		for c in self.components() {
-			if c.commands(&mut res, force_all)
-				!= CommandBlocking::PassingOn
-				&& !force_all
-			{
-				break;
-			}
-		}
+		command_pump(&mut res, force_all, &self.components());
 
 		res.push(CommandInfo::new(
 			strings::commands::find_file(&self.key_config),